# docs/swagger.yaml
basePath: /api
definitions:
...
  domain.Flexibility:
    enum:
    - 0
    - 30
    - 60
    type: integer
    x-enum-varnames:
    - Exact
    - Approximate
    - Flexible
  domain.Request:
    properties:
      Date:
        description: zero → anytime (no date constraint)
        type: string
      DepartureAt:
        description: zero → day or anytime (no time constraint)
        type: string
      Destination:
        type: string
      ExpiresAt:
        type: string
      Flexibility:
        allOf:
        - $ref: '#/definitions/domain.Flexibility'
        description: only meaningful when DepartureAt is set
      ID:
        type: string
      Origin:
        type: string
      Phone:
        type: string
      PostedAt:
        type: string
      SearcherName:
        type: string
    type: object
...
  handler.DeleteRequestBody:
    properties:
      phone:
        type: string
    type: object
...
  handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  handler.PostRequestBody:
    properties:
      departure_at:
        description: RFC3339 — specific time mode
        type: string
      departure_date:
        description: YYYY-MM-DD — day mode
        type: string
      departure_time:
        description: HH:MM — daily mode
        type: string
      destination:
        type: string
      flexibility:
        type: integer
      origin:
        type: string
      phone:
        type: string
      searcher_name:
        type: string
    type: object
...
info:
  contact: {}
  description: Local ride-sharing notice board API.
  title: Go-Stop API
  version: "1.0"
paths:
...
  /requests:
    get:
      operationId: listRequests
      parameters:
      - description: Searcher phone
        in: header
        name: X-Phone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Request'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      tags:
      - requests
    post:
      consumes:
      - application/json
      operationId: createRequest
      parameters:
      - description: Alert to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.PostRequestBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Request'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      tags:
      - requests
  /requests/{id}:
    delete:
      consumes:
      - application/json
      operationId: deleteRequest
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: string
      - description: Owner phone
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.DeleteRequestBody'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      tags:
      - requests
    get:
      operationId: getRequest
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: string
      - description: Owner phone
        in: header
        name: X-Phone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Request'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      tags:
      - requests
...
swagger: "2.0"